@@ -1,3 +1,3 @@
 module ssh-agent-switcher
 
-go 1.20
+go 1.22
